+++ src/app/page.tsx
@@ -1,13 +1,17 @@
-import Image from 'next/image'
-import styles from './page.module.css'
-import Header from './components/Header'
+import Image from "next/image";
+import styles from "./page.module.css";
+import React from "react";
+import Header from "../components/Header";
 
 export default function Home() {
   return (
-    <Header>
-    <main className={styles.main}>
-      <div className="header-title" background='black'><h1>Testes</h1></div>
+    <>
+      <Header />
+      <main className={styles.main}>
+        <div>
+          <h1>Testes</h1>
+        </div>
       </main>
-    </Header>
-  )
+    </>
+  );
 }
