+++ src/components/Header/styles.module.css
@@ -1,18 +1,15 @@
 .container {
-  flex: auto;
-  padding: 15px;
-  background: #ffee00;
-  color: rgb(185, 19, 19);
-  font-size: 20px;
+  padding: 12px;
+  background: #352c2c;
+  font-size: 1rem;
   text-align: center;
 }
 
 .searchButtonBox {
-  /* flex: auto; */
   display: flex;
   justify-content: space-between;
   align-items: center;
-  gap: 24px;
+  gap: 1.5rem;
 }
