+++ src/components/SearchIcon/styles.module.css
@@ -1,0 +1,9 @@
+.searchButtonBox {
+  /* flex: auto; */
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 24px;
+}
+
+
